Continue Debugging...
---
Now I need to make the apple not to take
a position of the blocks of the snake,
but I can't seem to find a way to do it.

@scripts/apple.js	Re-added the line that
cleared the Alarm of the apple in the disable()
method

@scripts/board.js	Now the board only gets
a random position in getRandomPosition() and
doesn't remove it from the free positions.
Also, used .indexOf in occupyPosition(num, num)
rather than .findIndex

@scripts/snake.js	Created execOnAll(cb),
a function that executes a callback on
the blocks and the head of the snake