Fixed Apple Bug
---
Now the apple doesn't generate in the same
block than some block of the snake.

I fixed it by changing completely the approach
I had; instead of giving the board awareness
of its positions and occupying and disoccupying
all the time, I decided to only get the
positions of the snake and then substract them
from the positions of the board, this way
getting the current free positions every
time the snake moves.

@scripts/board.js	Deleted all functions
related to position awareness (occupy,
disocuppy and isFree) and added a method
to substract give positions from the
positions of the board.

@scripts/main.js	Created an auxiliar
function `getRandomElement` to help
with the new approach for the apple.
Adapted the function `generateApple` and
created `generateSnake` to get a barely
customizable snake.
Also, deleted unnecessary comments.

@scripts/snake.js	Created a method to get
the positions of all the blocks of the snake
and also changed the constructor