Added Control Panel
---
Only the html skeleton; I will focus on the
css in the next days.

@scripts/main.js	I made a function
`roundToHalfs` that helps with the points
of the user, so that it only shows multiples
of 0.5 instead of taking up the whole screen.

Also, I need to make the restart button work
as well as the pause one (the alarm of the
apple doesn't actually pause; it continues
running and then the apple disappears).

Regarding to the apple, it also seems that the
it is being placed in spaces within the
board that don't show (maybe it goes to the
end of the grid), I should also fix that.