Added Minimalistic Design
---
I created the design for the head and block
of the snake in the minimalistic folder.

Metrics such as the fill color, stroke,
stroke width, etc... should be all customizable
in the future, but for now, I will just let them
to be green, orange, and 16, respectively.

All snake design files should have instructions
to replace their path points depending on the
stroke width, currently, the files I edited
have them.