Implemented SVGs
---
Now the snake and apple have inner svgs that
make them look more friendly.

It is intended that the player can
customize the svgs by using the menu,
I was thinking about creating a factory
of svg designs, using the text of the svg
files I've created so far.

@main.css	Modified the snake blocks (and
apple) to be the same color as the blocks
of the board.

Also, set `transition: translate ease` for
my future plans of animating the movement of
the blocks with css