Re-refactored Alarm Class
---
Turns out my previous approach could pause the
apple accurately only once, and this one can do it ilimited
times, which I guess is great.

I used the idea of "timemarkers" (dates that serve
as a reference); I use a variable to store the
time passed from the time marker (initially set to
the date the alarm started tu run for the first time)
to the moment the alarm is paused (to save the time the
alarm has been running).

Then, whenever the alarm is set to run again, I move
the marker to the current date (for next pauses,
to save the time the alarm has been running,
respect to the date it was set to run again).

The marker is also used as a reference for getting the left
time and ellapsed time of the alarm.

Finally, I deleted the variables start and end variables of
the alarm because it was very unecessary.