Fixed Apple Bug
---
Turns out the problem wasn't the blocks pushing each
other (that doesn't happen once they get their column
and row set) but the positions the board had.

When I looped to create the positions, I used zero-based
indeces that then I pushed to the positions array.

The grid column/row property uses one-based indeces, so,
sometimes when I grabbed a position from the board, the
apple got a 0 coordinate, which was invalid and changed
to `auto` at the moment of setting the column/row.

Finally, the `auto` value simply made the apple go
at the end of the grid, causing it to overflow.

How I fixed it? Adding 1 to the coordinates pushed to
the positions array.

Also, I made the board calculate the length of the
blocks inside it so that I don't have to manipulate
css in the future and the board remains the same
size no matter its dimension.